# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BudgetIn": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "kind": {
            "$ref": "#/components/schemas/BudgetKind",
            "default": "all"
          },
          "limit_paise": {
            "exclusiveMinimum": 0.0,
            "title": "Limit Paise",
            "type": "integer"
          },
          "name": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/BudgetPeriod"
          },
          "rollover": {
            "default": false,
            "title": "Rollover",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "limit_paise",
          "period"
        ],
        "title": "BudgetIn",
        "type": "object"
      },
      "BudgetKind": {
        "enum": [
          "all",
          "added"
        ],
        "title": "BudgetKind",
        "type": "string"
      },
      "BudgetOut": {
        "additionalProperties": false,
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/BudgetKind"
          },
          "limit_paise": {
            "exclusiveMinimum": 0.0,
            "title": "Limit Paise",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/BudgetPeriod"
          },
          "rollover": {
            "title": "Rollover",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "category_id",
          "name",
          "limit_paise",
          "period",
          "kind",
          "rollover",
          "archived",
          "created_at",
          "updated_at"
        ],
        "title": "BudgetOut",
        "type": "object"
      },
      "BudgetPatch": {
        "additionalProperties": false,
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "limit_paise": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Paise"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 60,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "rollover": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rollover"
          }
        },
        "title": "BudgetPatch",
        "type": "object"
      },
      "BudgetPeriod": {
        "enum": [
          "month",
          "fy",
          "custom"
        ],
        "title": "BudgetPeriod",
        "type": "string"
      },
      "BudgetStatus": {
        "enum": [
          "on_pace",
          "projected_over",
          "over",
          "pending"
        ],
        "title": "BudgetStatus",
        "type": "string"
      },
      "BudgetSuggestion": {
        "additionalProperties": false,
        "properties": {
          "average_spent_paise": {
            "title": "Average Spent Paise",
            "type": "integer"
          },
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "months": {
            "title": "Months",
            "type": "integer"
          },
          "suggested_limit_paise": {
            "exclusiveMinimum": 0.0,
            "title": "Suggested Limit Paise",
            "type": "integer"
          }
        },
        "required": [
          "category_id",
          "suggested_limit_paise",
          "average_spent_paise",
          "months"
        ],
        "title": "BudgetSuggestion",
        "type": "object"
      },
      "BudgetTrail": {
        "additionalProperties": false,
        "description": "The evidence behind one budget row: what the last months actually cost (the\nsparkline), whether the line has been held and for how long, and the daily\ncumulative climb against an even pace.",
        "properties": {
          "budget_id": {
            "title": "Budget Id",
            "type": "string"
          },
          "daily_cumulative_paise": {
            "items": {
              "type": "integer"
            },
            "title": "Daily Cumulative Paise",
            "type": "array"
          },
          "even_pace_paise": {
            "items": {
              "type": "integer"
            },
            "title": "Even Pace Paise",
            "type": "array"
          },
          "held_months_running": {
            "title": "Held Months Running",
            "type": "integer"
          },
          "limit_paise": {
            "title": "Limit Paise",
            "type": "integer"
          },
          "months": {
            "items": {
              "$ref": "#/components/schemas/MonthSpend"
            },
            "title": "Months",
            "type": "array"
          }
        },
        "required": [
          "budget_id",
          "limit_paise",
          "months",
          "held_months_running",
          "daily_cumulative_paise",
          "even_pace_paise"
        ],
        "title": "BudgetTrail",
        "type": "object"
      },
      "BudgetView": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/BudgetOut"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "pace": {
            "$ref": "#/components/schemas/PaceOut"
          },
          "window_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          }
        },
        "required": [
          "budget",
          "category",
          "window_start",
          "window_end",
          "pace"
        ],
        "title": "BudgetView",
        "type": "object"
      },
...
      "CategoryKind": {
        "enum": [
          "expense",
          "income"
        ],
        "title": "CategoryKind",
        "type": "string"
      },
      "CategoryOut": {
        "additionalProperties": false,
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/CategoryKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "icon",
          "kind",
          "sort_order",
          "archived",
          "created_at",
          "updated_at"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonthSpend": {
        "additionalProperties": false,
        "properties": {
          "held": {
            "title": "Held",
            "type": "boolean"
          },
          "month": {
            "title": "Month",
            "type": "string"
          },
          "spent_paise": {
            "title": "Spent Paise",
            "type": "integer"
          }
        },
        "required": [
          "month",
          "spent_paise",
          "held"
        ],
        "title": "MonthSpend",
        "type": "object"
      },
...
      "PaceOut": {
        "additionalProperties": false,
        "properties": {
          "elapsed_days": {
            "title": "Elapsed Days",
            "type": "integer"
          },
          "fraction_elapsed": {
            "title": "Fraction Elapsed",
            "type": "number"
          },
          "fraction_spent": {
            "title": "Fraction Spent",
            "type": "number"
          },
          "limit_paise": {
            "title": "Limit Paise",
            "type": "integer"
          },
          "projected_overspend_paise": {
            "title": "Projected Overspend Paise",
            "type": "integer"
          },
          "projected_paise": {
            "title": "Projected Paise",
            "type": "integer"
          },
          "remaining_paise": {
            "title": "Remaining Paise",
            "type": "integer"
          },
          "spent_paise": {
            "title": "Spent Paise",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/BudgetStatus"
          },
          "total_days": {
            "title": "Total Days",
            "type": "integer"
          },
          "upcoming_paise": {
            "title": "Upcoming Paise",
            "type": "integer"
          }
        },
        "required": [
          "spent_paise",
          "limit_paise",
          "elapsed_days",
          "total_days",
          "upcoming_paise",
          "remaining_paise",
          "fraction_spent",
          "fraction_elapsed",
          "projected_paise",
          "status",
          "projected_overspend_paise"
        ],
        "title": "PaceOut",
        "type": "object"
      },
...
      "RebalanceIn": {
        "additionalProperties": false,
        "properties": {
          "budget_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 2,
            "title": "Budget Ids",
            "type": "array"
          }
        },
        "required": [
          "budget_ids"
        ],
        "title": "RebalanceIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BudgetBox",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/budgets": {
      "get": {
        "operationId": "list_budgets_v1_budgets_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetOut"
                  },
                  "title": "Response List Budgets V1 Budgets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Budgets",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/pace": {
      "get": {
        "operationId": "pace_v1_budgets_pace_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetView"
                  },
                  "title": "Response Pace V1 Budgets Pace Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pace",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/rebalance": {
      "post": {
        "operationId": "rebalance_v1_budgets_rebalance_post",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebalanceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetOut"
                  },
                  "title": "Response Rebalance V1 Budgets Rebalance Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rebalance",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/suggestions": {
      "get": {
        "operationId": "suggestions_v1_budgets_suggestions_get",
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 24,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetSuggestion"
                  },
                  "title": "Response Suggestions V1 Budgets Suggestions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Suggestions",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/{budget_id}": {
      "patch": {
        "operationId": "patch_budget_v1_budgets__budget_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Budget",
        "tags": [
          "budgets"
        ]
      },
      "put": {
        "operationId": "upsert_budget_v1_budgets__budget_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Budget",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/{budget_id}/trail": {
      "get": {
        "description": "The sparkline, the 'held its line N months running' streak, and this\nperiod's daily climb against an even pace \u2014 the budget row's whole case.",
        "operationId": "trail_v1_budgets__budget_id__trail_get",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 24,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetTrail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trail",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/budgets/{budget_id}/txns/{txn_id}": {
      "delete": {
        "operationId": "remove_txn_v1_budgets__budget_id__txns__txn_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Txn",
        "tags": [
          "budgets"
        ]
      },
      "put": {
        "operationId": "add_txn_v1_budgets__budget_id__txns__txn_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Txn",
        "tags": [
          "budgets"
        ]
      }
    },
...
  }
}